extender feature flag feature to be able to select multiple flags
