+++ cypress/e2e/wallet-dropdown.test.ts
@@ -1,17 +1,20 @@
 import { getTestSelector } from '../utils'
+import { selectFeatureFlag } from '../utils'
+import { FeatureFlag } from '../../src/featureFlags/flags/featureFlags'
 
 describe('Wallet Dropdown', () => {
   before(() => {
     cy.visit('/')
+    selectFeatureFlag([FeatureFlag.navBar, FeatureFlag.tokens, FeatureFlag.tokenSafety])
   })
 
-  it.skip('should change the theme', () => {
+  it('should change the theme', () => {
     cy.get(getTestSelector('navbar-wallet-dropdown')).click()
     cy.get(getTestSelector('wallet-select-theme')).click()
     cy.get(getTestSelector('wallet-select-theme')).contains('Light theme').should('exist')
   })
 
-  it.skip('should select a language', () => {
+  it('should select a language', () => {
     cy.get(getTestSelector('wallet-select-language')).click()
     cy.get(getTestSelector('wallet-language-item')).contains('Afrikaans').click({ force: true })
     cy.get(getTestSelector('wallet-header')).should('contain', 'Taal')
@@ -20,13 +23,13 @@ describe('Wallet Dropdown', () => {
     cy.get(getTestSelector('wallet-back')).click()
   })
 
-  it.skip('should open the wallet connect modal from the drop down', () => {
+  it('should open the wallet connect modal from the drop down', () => {
     cy.get(getTestSelector('wallet-connect-wallet')).click()
     cy.get(getTestSelector('wallet-modal')).should('exist')
     cy.get(getTestSelector('wallet-modal-close')).click()
   })
 
-  it.skip('should open the wallet connect modal from the navbar', () => {
+  it('should open the wallet connect modal from the navbar', () => {
     cy.get(getTestSelector('navbar-connect-wallet')).click()
     cy.get(getTestSelector('wallet-modal')).should('exist')
     cy.get(getTestSelector('wallet-modal-close')).click()
