+++ src/featureFlags/flags/featureFlags.ts
@@ -1,0 +1,8 @@
+export enum FeatureFlag {
+  navBar = 'navBar',
+  nft = 'nfts',
+  redesign = 'redesign',
+  tokens = 'tokens',
+  tokensNetworkFilter = 'tokensNetworkFilter',
+  tokenSafety = 'tokenSafety',
+}
